update: just backing work up

+++ includes/packet.h
@@ -33,7 +33,7 @@ void logPack(pack *input) {
 	dbg(LOG_CHANNEL, "Package [Src: %hhu, Dest: %hhu, Seq: %hhu, TTL: %hhu, Protocol: %hhu, Payload: %s]\n", input->src, input->dest, input->seq, input->TTL, input->protocol, input->payload);
 }
 
-void makePack(pack *Package, uint16_t src, uint16_t dest, uint16_t TTL, uint16_t protocol, uint16_t seq, uint8_t* payload, uint8_t length) {
+void makePack(pack *Package, uint16_t src, uint16_t dest, uint8_t TTL, uint8_t protocol, uint16_t seq, uint8_t* payload, uint8_t length) {
 	Package->src = src;
 	Package->dest = dest;
 	Package->TTL = TTL;
